Fixed a minor issue in docker_backup_tests.sh Added a check if the file exists before deleting it

+++ backup_tests/docker_backup_tests.sh
@@ -221,7 +221,9 @@ test_pxb_docker() {
 }
 
 # Check and clean existing installation
-rm backup_log
+if [ -f backup_log ]; then
+    rm backup_log
+fi
 clean_setup
 test_pxb_docker | tee -a backup_log
